requirements.txt
streamlit
python-dotenv
swarmauri
typer<0.10.0
spacy
weasel
gradio==4.43.0
fastapi==0.112.4	
gradio-client==1.3.0
pydantic==2.9.0
pydantic-core==2.23.2
